+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.8.2 (Unreleased)
+
+IMPROVEMENTS:
+ * client: Create new process group on process startup. [[GH-3572](https://github.com/hashicorp/nomad/issues/3572)]
+
+BUG FIXES:
+ * driver/exec: Create process group for Windows process and send Ctrl-Break signal on Shutdown [[GH-2117](https://github.com/hashicorp/nomad/issues/2117)]
+
 ## 0.8.1 (April 17, 2018)
 
 BUG FIXES: